Update index.js
The new native BigInt datatype supports parsing from a string type.  I added string as an allowable type before attempting to parse, then wrapping the parsing with a try/catch in case the string is invalid.  I did this because while my API now supports BigInt with this library, the front-end portion connecting to the API does not and sends the value as a string.